requirements.apt -> requirements-docker.apt
These are dependencies meant to be installed into the docker container,
but are not actually dependencies of the python package that might
eventually exist here.

In particular, it would be silly to assume that bitcoind is the only
compatible bitcoin implementation that this works with. So it's not a
real dependency, just something that needs to get installed into the
container for development and demonstration purposes....